requirements.txt
Flask==3.0.3
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
torch==2.3.1
transformers==4.42.3